requirements.txt
PyMySQL==0.7.11
lxml==3.8.0
pycrypto==2.6.1
Pillow==4.3.0
web==0.6.0
